# vectorized StandardParams linear_ramp_from_hamiltonian