Replaced the for loops in calculation of betas and gammas with a linspace version, resulting in roughly 10x faster execution of StandardParams linear_ramp_from_hamiltonian method as used in example notebook #6. 

Runtime before optimization: 1.22 ms ± 10.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

Runtime after optimization: 85.9 µs ± 905 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)