@jlbosse 2019-10-31
Hey @ysinha1 ,
Thank you for the pull request, the code changes look all very sensible. We are currently discussing, how to best merge them into our code base, since most of our development is happening in a different private repo, but will definitely merge it! Thank you for your contribution.

About the failing checks: That really has more to do with our tests setup wrongly, than with you code. So I guess, thanks for pointing that out as well :D
Cheers, Jan Lukas