@ysinha1 2019-11-04
> Hey @ysinha1 ,
> Thank you for the pull request, the code changes look all very sensible. We are currently discussing, how to best merge them into our code base, since most of our development is happening in a different private repo, but will definitely merge it! Thank you for your contribution.
> 
> About the failing checks: That really has more to do with our tests setup wrongly, than with you code. So I guess, thanks for pointing that out as well :D
> Cheers, Jan Lukas

Thanks for the feedback @jlbosse ! Looking forward to seeing the latest version once it's public. Cheers.